Initial commit
Run the init function when the page is fully loaded. Sets up the event handlers for the page.  the purpose of the function is to change what appears in the cal window when the user clicks the buttons. Getting the value from the text area box with the id of calcWindow and for decimals.  defines each button on the calculator. Putting the text on the window of the calculator. Put the cursor focus with in the calculator window. Keeps the browser from preforming the default action in response to the user pressing the up-arrow key. Putting the text on the window of the calculator.